+++ ChangeLog.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 
 
-## [0.7.3] - 2018-08-08
+## [0.8.0] - 2018-08-09
 
 ### Added